requirements.txt
fastapi
uvicorn 
requests
python-dotenv
azure-search-documents
azure-storage-blob
sentence-transformers